refactor createProfile to a command file
